+++ src/components/data/static/loanActionMap.js
@@ -2,6 +2,7 @@ export default {
   'checkedout': 'ui-users.data.loanActionMap.checkedOut',
   'checkedin': 'ui-users.data.loanActionMap.checkedIn',
   'declaredLost': 'ui-users.data.loanActionMap.declaredLost',
+  'claimedReturned': 'ui-users.data.loanActionMap.claimedReturned',
   'renewed': 'ui-users.data.loanActionMap.renewed',
   'Operator': 'ui-users.data.loanActionMap.source',
   'holdrequested': 'ui-users.data.loanActionMap.holdRequested',