+++ test/bigtest/tests/closed-loans-test.js
@@ -5,6 +5,7 @@ import {
 } from '@bigtest/mocha';
 import { expect } from 'chai';
 
+import translations from '../../../translations/ui-users/en';
 import setupApplication from '../helpers/setup-application';
 import ClosedLoansInteractor from '../interactors/closed-loans';
 import UsersInteractor from '../interactors/users';
@@ -29,8 +30,19 @@ describe('Closed Loans', () => {
       user = this.server.create('user');
       const loans = this.server.createList('loan', 5, 'feesAndFines', {
         userId: user.id,
-        item: (i) => ({ ...this.item, barcode: i }),
-        status: { name: 'Closed' }
+        item: (i) => ({
+          ...this.item,
+          barcode: i,
+          callNumberComponents: {
+            prefix: 'prefix',
+            callNumber: 'callNumber',
+            suffix: 'suffix',
+          },
+          enumeration: 'enumeration',
+          chronology: 'chronology',
+          volume: 'volume',
+        }),
+        status: { name: 'Closed' },
       });
 
       setupAnonymizationAPIResponse(this.server, [{
@@ -41,7 +53,7 @@ describe('Closed Loans', () => {
         }]
       }]);
 
-      this.visit(`/users/${loans[0].userId}/loans/closed`);
+      this.visit(`/users/${loans[0].userId}/loans/closed?query=%20&sort=requests`);
 
       await ClosedLoansInteractor.whenLoaded();
     });
@@ -57,6 +69,16 @@ describe('Closed Loans', () => {
     describe('loan list', () => {
       it('should be presented', () => {
         expect(ClosedLoansInteractor.list.isPresent).to.be.true;
+      });
+
+      it('loan should have effective call number', () => {
+        const callNumber = 'prefix callNumber suffix volume enumeration chronology';
+
+        expect(ClosedLoansInteractor.callNumbers(0).text).to.equal(callNumber);
+      });
+
+      it('loan table call number column header should have title', () => {
+        expect(ClosedLoansInteractor.columnHeaders(4).text).to.equal(translations['loans.details.effectiveCallNumber']);
       });
     });